+++ seacatauth/openidconnect/session.py
@@ -12,8 +12,8 @@ def oauth2_session_builder(oauth2_data):
 	if scope is not None:
 		scope = list(scope)
 
-	yield (SessionAdapter.FNOAuth2Scope, scope)
-	yield (SessionAdapter.FNOAuth2ClientId, oauth2_data["client_id"])
-	yield (SessionAdapter.FNOAuth2AccessToken, secrets.token_bytes(token_length))
-	yield (SessionAdapter.FNOAuth2RefreshToken, secrets.token_bytes(token_length))
-	yield (SessionAdapter.FNOAuth2IdToken, secrets.token_bytes(token_length))
+	yield (SessionAdapter.FN.OAuth2.Scope, scope)
+	yield (SessionAdapter.FN.OAuth2.ClientId, oauth2_data["client_id"])
+	yield (SessionAdapter.FN.OAuth2.AccessToken, secrets.token_bytes(token_length))
+	yield (SessionAdapter.FN.OAuth2.RefreshToken, secrets.token_bytes(token_length))
+	yield (SessionAdapter.FN.OAuth2.IdToken, secrets.token_bytes(token_length))