+++ CHANGELOG.md
@@ -4,10 +4,22 @@
 
 ### Fix
 - Use datetime.datetime.utcnow (#29, PLUM Sprint 220422)
+- After-provisioning cleanup (#33, PLUM Sprint 220506)
+- Fix session expiration for back-compat (#34, PLUM Sprint 220506)
+- ~~Import ASAB web container (#36, PLUM Sprint 220520)~~(4ade2c60)
+- Store login session data in database (#37, PLUM Sprint 220520)
+
+### Features
+- Mock mode in SMS provider (#37, PLUM Sprint 220520)
 
 ### Refactoring
 - Persistent OIDC authorization codes (#25, PLUM Sprint 220408)
 - Persistent TOTP secrets (#27, PLUM Sprint 220408)
+- Use two-stage docker build (#31, PLUM Sprint 220520)
+- Revise default configs and examples (#28, PLUM Sprint 220422)
+- Persistent login sessions (#26, PLUM Sprint 220422)
+- Session adapter restructured (#32, PLUM Sprint 220506)
+- Credentials deletion via credentials service (#33, PLUM Sprint 220506)
 
 ---